Simplify LogoList component
